+++ src/NServiceBus.Core/NServiceBus.Core.csproj
@@ -136,6 +136,7 @@
     <Compile Include="Persistence\Development\SagaPersister\DevelopmentSagaIdGenerator.cs" />
     <Compile Include="Persistence\Development\SagaPersister\DevelopmentSagaPersistence.cs" />
     <Compile Include="Persistence\Development\SagaPersister\DevelopmentSagaPersister.cs" />
+    <Compile Include="Persistence\Development\SagaPersister\DevelopmentStorageAdapter.cs" />
     <Compile Include="Persistence\Development\SagaPersister\DevelopmentSyncronizedStorage.cs" />
     <Compile Include="Persistence\Development\SagaPersister\DevelopmentSyncronizedStorageSession.cs" />
     <Compile Include="Persistence\Development\SagaPersister\SagaManifest.cs" />
