+++ src/NServiceBus.Core/Persistence/Development/SagaPersister/DevelopmentSagaPersistence.cs
@@ -2,12 +2,7 @@
 {
     using System;
     using System.IO;
-    using System.Threading.Tasks;
-    using Extensibility;
-    using NServiceBus.Outbox;
     using NServiceBus.Sagas;
-    using Persistence;
-    using Transport;
 
     class DevelopmentSagaPersistence : Feature
     {
@@ -26,28 +21,15 @@
             var storageLocation = context.Settings.Get<string>(StorageLocationKey);
 
             var allSagas = context.Settings.Get<SagaMetadataCollection>();
-            
+
             var sagaManifests = new SagaManifestCollection(allSagas,storageLocation);
 
-            context.Container.ConfigureComponent<DevelopmentSyncronizedStorage>(DependencyLifecycle.SingleInstance);
+            context.Container.ConfigureComponent(b=>new DevelopmentSyncronizedStorage(sagaManifests),  DependencyLifecycle.SingleInstance);
             context.Container.ConfigureComponent<DevelopmentStorageAdapter>(DependencyLifecycle.SingleInstance);
 
-            context.Container.ConfigureComponent(b => new DevelopmentSagaPersister(sagaManifests), DependencyLifecycle.SingleInstance);
+            context.Container.ConfigureComponent(b => new DevelopmentSagaPersister(), DependencyLifecycle.SingleInstance);
         }
 
         internal static string StorageLocationKey = "DevelopmentSagaPersistence.StorageLocation";
-    }
-
-    class DevelopmentStorageAdapter : ISynchronizedStorageAdapter
-    {
-        public Task<CompletableSynchronizedStorageSession> TryAdapt(OutboxTransaction transaction, ContextBag context)
-        {
-            return Task.FromResult<CompletableSynchronizedStorageSession>(null);
-        }
-
-        public Task<CompletableSynchronizedStorageSession> TryAdapt(TransportTransaction transportTransaction, ContextBag context)
-        {
-            return Task.FromResult<CompletableSynchronizedStorageSession>(null);
-        }
     }
 }
